Updates + Code Restructure
