+++ setup/requirements_tensorflow_apple_silicon.txt
@@ -1,0 +1,7 @@
+tensorflow-macos==2.12.0
+tensorflow-metal==0.8.0
+numpy==1.23.5
+ipykernel==6.22.0  # needed to run jupyter notebooks 
+tensorboard==2.12.3
+matplotlib==3.7.1
+nbconvert==7.3.1  # needed for converting ipynb to other formats such as python, html and pdf
